Rename VSLSPTagHelperTooltipFactory to ClassifiedTagHelperTooltipFactory
