+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Completion/RazorCompletionItemResolverTest.cs
@@ -20,7 +20,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion;
 public class RazorCompletionItemResolverTest : LanguageServerTestBase
 {
     private readonly MarkupTagHelperTooltipFactory _markupTagHelperTooltipFactory;
-    private readonly VSLSPTagHelperTooltipFactory _vsLspTagHelperTooltipFactory;
+    private readonly ClassifiedTagHelperTooltipFactory _classifiedTagHelperTooltipFactory;
     private readonly VSInternalCompletionSetting _completionCapability;
     private readonly VSInternalClientCapabilities _defaultClientCapability;
     private readonly VSInternalClientCapabilities _vsClientCapability;
@@ -33,7 +33,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
         var projectManager = CreateProjectSnapshotManager();
 
         _markupTagHelperTooltipFactory = new MarkupTagHelperTooltipFactory(projectManager);
-        _vsLspTagHelperTooltipFactory = new DefaultVSLSPTagHelperTooltipFactory(projectManager);
+        _classifiedTagHelperTooltipFactory = new DefaultClassifiedTagHelperTooltipFactory(projectManager);
         _completionCapability = new VSInternalCompletionSetting()
         {
             CompletionItem = new CompletionItemSetting()
@@ -69,7 +69,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_DirectiveCompletion_ReturnsCompletionItemWithDocumentation()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.Directive);
         razorCompletionItem.SetDirectiveCompletionDescription(new DirectiveCompletionDescription("Test directive"));
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -87,7 +87,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_MarkupTransitionCompletion_ReturnsCompletionItemWithDocumentation()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("@...", "@", RazorCompletionItemKind.MarkupTransition);
         razorCompletionItem.SetMarkupTransitionCompletionDescription(new MarkupTransitionCompletionDescription("Test description"));
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -105,7 +105,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_DirectiveAttributeCompletion_ReturnsCompletionItemWithDocumentation()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.DirectiveAttribute);
         razorCompletionItem.SetAttributeCompletionDescription(_attributeDescription);
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -123,7 +123,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_DirectiveAttributeParameterCompletion_ReturnsCompletionItemWithDocumentation()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.DirectiveAttributeParameter);
         razorCompletionItem.SetAttributeCompletionDescription(_attributeDescription);
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -141,7 +141,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_TagHelperElementCompletion_ReturnsCompletionItemWithDocumentation()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.TagHelperElement);
         razorCompletionItem.SetTagHelperElementDescriptionInfo(_elementDescription);
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -159,7 +159,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_TagHelperAttribute_ReturnsCompletionItemWithDocumentation()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.TagHelperAttribute);
         razorCompletionItem.SetAttributeCompletionDescription(_attributeDescription);
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -177,7 +177,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_VS_DirectiveAttributeCompletion_ReturnsCompletionItemWithDescription()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.DirectiveAttribute);
         razorCompletionItem.SetAttributeCompletionDescription(_attributeDescription);
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -195,7 +195,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_VS_DirectiveAttributeParameterCompletion_ReturnsCompletionItemWithDescription()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.DirectiveAttributeParameter);
         razorCompletionItem.SetAttributeCompletionDescription(_attributeDescription);
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -213,7 +213,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_VS_TagHelperElementCompletion_ReturnsCompletionItemWithDescription()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.TagHelperElement);
         razorCompletionItem.SetTagHelperElementDescriptionInfo(_elementDescription);
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -231,7 +231,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_VS_TagHelperAttribute_ReturnsCompletionItemWithDescription()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.TagHelperAttribute);
         razorCompletionItem.SetAttributeCompletionDescription(_attributeDescription);
         var completionList = CreateLSPCompletionList(razorCompletionItem);
@@ -249,7 +249,7 @@ public class RazorCompletionItemResolverTest : LanguageServerTestBase
     public async Task ResolveAsync_NonTagHelperCompletion_Noops()
     {
         // Arrange
-        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _vsLspTagHelperTooltipFactory);
+        var resolver = new RazorCompletionItemResolver(_markupTagHelperTooltipFactory, _classifiedTagHelperTooltipFactory);
         var completionItem = new VSInternalCompletionItem();
         var completionList = new VSInternalCompletionList() { Items = new[] { completionItem } };
